Accelerate tests and make ES in that case reliableby letting index build be explicetely if needed
