+++ conftest.py
@@ -12,15 +12,20 @@ from tempfile import NamedTemporaryFile
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.core.files.uploadedfile import InMemoryUploadedFile
+from django.core.management import call_command
 from django.urls import reverse
 
 from deposit.models import Repository
 from dissemin.settings import BASE_DIR
 from dissemin.settings import POSSIBLE_LANGUAGE_CODES
 from papers.baremodels import PAPER_TYPE_CHOICES
+from papers.models import Department
+from papers.models import Institution
+from papers.models import Name
 from papers.models import Paper
 from papers.models import OaiRecord
 from papers.models import OaiSource
+from papers.models import Researcher
 from publishers.models import Journal
 from publishers.models import Publisher
 
@@ -516,3 +521,38 @@ class LoadOaiSource():
         Provides BASE OaiSource. It is in the database from a migration. We do not add it to the list of to be deleted OaiSources
         """
         return OaiSource.objects.get(identifier='base')
+
+
+# Fixtures and Functions for load_test_data, which should prefereably not be used as it loads a lot of things
+def get_researcher_by_name(first, last):
+    n = Name.lookup_name((first, last))
+    return Researcher.objects.get(name=n)
+
+
+@pytest.fixture
+def load_test_data(request, db, django_db_setup, django_db_blocker):
+    with django_db_blocker.unblock():
+        call_command('loaddata', 'test_dump.json')
+        self = request.cls
+        self.i = Institution.objects.get(name='ENS')
+        self.d = Department.objects.get(name='Chemistry dept')
+        self.di = Department.objects.get(name='Comp sci dept')
+
+        self.r1 = get_researcher_by_name('Isabelle', 'Aujard')
+        self.r2 = get_researcher_by_name('Ludovic', 'Jullien')
+        self.r3 = get_researcher_by_name('Antoine', 'Amarilli')
+        self.r4 = get_researcher_by_name('Antonin', 'Delpeuch')
+        self.r5 = get_researcher_by_name('Terence', 'Tao')
+        self.hal = OaiSource.objects.get(identifier='hal')
+        self.arxiv = OaiSource.objects.get(identifier='arxiv')
+        self.lncs = Journal.objects.get(issn='0302-9743')
+        self.acm = Journal.objects.get(issn='1529-3785').publisher
+
+
+@pytest.fixture
+def rebuild_index(request):
+    rebuild_index = (
+        lambda: call_command('rebuild_index', interactive=False)
+    )
+    rebuild_index()
+    request.addfinalizer(rebuild_index)
