--- papers/conftest.py
@@ -1,8 +1,0 @@
-
-from backend.conftest import (
-    load_test_data as papers_load_test_data,
-    rebuild_index as papers_rebuild_index
-)
-
-load_test_data = papers_load_test_data
-rebuild_index = papers_rebuild_index
