--- publishers/conftest.py
@@ -1,4 +1,0 @@
-
-from backend.conftest import load_test_data as papers_load_test_data
-
-load_test_data = papers_load_test_data
